+++ srcs/ft_printf/ft_str_conv.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_str_conv.c                                      :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: guvillat <marvin@42.fr>                    +#+  +:+       +#+        */
+/*   By: arsciand <arsciand@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/01/23 15:27:43 by guvillat          #+#    #+#             */
-/*   Updated: 2020/03/02 16:50:29 by mpivet-p         ###   ########.fr       */
+/*   Updated: 2020/03/05 00:17:05 by arsciand         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,9 +26,9 @@ static int		wstring_handler(PF *argument, va_list ap)
 	argument->spec = 'S';
 	if (!argument->warg)
 		argument->warg = L"(null)";
-	len = ft_wbytelen(argument->warg);
+	len = (ssize_t)ft_wbytelen(argument->warg);
 	if (argument->flags[0] > -1 && argument->flags[0] < len)
-		argument->arg = ft_wstrsub(argument->warg, 0, argument->flags[0]);
+		argument->arg = ft_wstrsub(argument->warg, 0, (size_t)argument->flags[0]);
 	else
 		argument->arg = ft_transform_wchar_in_char(argument->warg);
 	return (ft_print_str(argument));
@@ -46,9 +46,9 @@ int				string_handler(PF *argument, va_list ap)
 		argument->arg = ft_strdup("(null)");
 	else
 	{
-		len = ft_strlen(tmp);
+		len = (ssize_t)ft_strlen(tmp);
 		if (argument->flags[0] > -1 && argument->flags[0] < len)
-			argument->arg = ft_strsub(tmp, 0, argument->flags[0]);
+			argument->arg = ft_strsub(tmp, 0, (size_t)argument->flags[0]);
 		else
 			argument->arg = ft_strdup(tmp);
 	}
