+++ srcs/ft_putnbr_fd.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_putnbr_fd.c                                     :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: fcatusse <marvin@42.fr>                    +#+  +:+       +#+        */
+/*   By: arsciand <arsciand@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/11/06 16:06:12 by fcatusse          #+#    #+#             */
-/*   Updated: 2019/07/19 12:38:57 by fcatusse         ###   ########.fr       */
+/*   Updated: 2020/03/04 23:08:15 by arsciand         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,6 +23,6 @@ void	ft_putnbr_fd(int n, int fd)
 		nb = -nb;
 	}
 	if (nb > 9)
-		ft_putnbr_fd(nb / 10, fd);
+		ft_putnbr_fd((int)(nb / 10), fd);
 	ft_putchar_fd(nb % 10 + '0', fd);
 }
